requirements.txt
torch>=2.0
torchvision
numpy
pillow
scipy
datasets
wandb
